requirements_images.txt
Pillow>=9.0.0
numpy>=1.21.0


